[POC] im_livechat, *: remove Promise.reject in (widget) LivechatButton willStart lifecycle
*: website_livechat
