+++ addons/im_livechat/static/src/services/public_livechat_service.js
@@ -8,9 +8,12 @@ export const publicLivechatService = {
     dependencies: ['messaging'],
     async start(env, { messaging: messagingService }) {
         const messaging = await messagingService.get();
-        if (messaging.publicLivechatGlobal.livechatButtonView) {
-            const livechatButton = new LivechatButton(rootWidget, messaging);
-            livechatButton.appendTo(document.body);
-        }
+        return {
+            mountLivechatButton() {
+                const livechatButton = new LivechatButton(rootWidget, messaging);
+                livechatButton.appendTo(document.body);
+                return livechatButton;
+            },
+        };
     },
 };
