+++ addons/website_livechat/static/src/models/livechat_button_view.js
@@ -16,39 +16,17 @@ patchRecordMethods('LivechatButtonView', {
      * @override
      */
     async onChatbotRestartScript(ev) {
-        if (this.isWebsiteLivechatChatbotFlow) {
+        if (this.messaging.publicLivechatGlobal.isWebsiteLivechatChatbotFlow) {
             this.messaging.publicLivechatGlobal.update({ messages: clear() });
             this.widget._renderMessages();
         }
         return this._super(ev);
     },
     /**
-     * Check if a chat request is opened for this visitor
-     * if yes, replace the session cookie and start the conversation immediately.
-     * Do this before calling super to have everything ready before executing existing start logic.
-     * This is used for chat request mechanism, when an operator send a chat request
-     * from backend to a website visitor.
-     *
-     * @override
-     */
-    willStart() {
-        if (this.isTestChatbot) {
-            /**
-             * Overridden to avoid calling the "init" endpoint as it requires a im_livechat.channel linked
-             * to work properly.
-             */
-            return this.messaging.publicLivechatGlobal.loadQWebTemplate();
-        }
-        if (this.messaging.publicLivechatGlobal.options.chat_request_session) {
-            set_cookie('im_livechat_session', JSON.stringify(this.messaging.publicLivechatGlobal.options.chat_request_session), 60 * 60);
-        }
-        return this._super();
-    },
-    /**
      * @override
      */
     _computeIsOpenChatDebounced() {
-        if (this.isTestChatbot) {
+        if (this.messaging.publicLivechatGlobal.isTestChatbot) {
             return false;
         }
         return this._super();
@@ -63,7 +41,7 @@ patchRecordMethods('LivechatButtonView', {
      * @override
      */
     async _openChat() {
-        if (!this.isTestChatbot) {
+        if (!this.messaging.publicLivechatGlobal.isTestChatbot) {
             return this._super();
         }
         this.messaging.publicLivechatGlobal.update({
